+++ packages/common/src/index.ts
@@ -10,8 +10,6 @@ export * from "./util";
 export * from "./ide/util/messages";
 export { getKey, splitKey } from "./util/splitKey";
 export { hrtimeBigintToSeconds } from "./util/timeUtils";
-export * from "./util/walkSync";
-export * from "./util/walkAsync";
 export * from "./util/disposableFrom";
 export * from "./util/camelCaseToAllDown";
 export { Notifier } from "./util/Notifier";
@@ -46,11 +44,14 @@ export * from "./types/TextEditorDecorationType";
 export * from "./types/TextEditorEdit";
 export * from "./types/TextEditorOptions";
 export * from "./types/TextLine";
+export * from "./types/ScopeTestRecorderStorage";
 export * from "./types/Token";
 export * from "./types/HatTokenMap";
 export * from "./types/ScopeProvider";
 export * from "./types/SpokenForm";
 export * from "./types/commandHistory";
+export * from "./types/TalonSpokenForms";
+export * from "./types/TestHelpers";
 export * from "./util/textFormatters";
 export * from "./util/serializedMarksToTokenHats";
 export * from "./types/snippet.types";
@@ -63,16 +64,12 @@ export * from "./util/omitByDeep";
 export * from "./util/range";
 export * from "./util/uniqWithHash";
 export * from "./testUtil/testConstants";
-export * from "./testUtil/getFixturePaths";
-export * from "./testUtil/getCursorlessRepoRoot";
-export * from "./testUtil/runRecordedTest";
 export * from "./testUtil/serialize";
 export * from "./testUtil/shouldUpdateFixtures";
 export * from "./testUtil/TestCaseSnapshot";
 export * from "./testUtil/serializeTestFixture";
 export * from "./testUtil/getSnapshotForComparison";
 export * from "./testUtil/asyncSafety";
-export * from "./testUtil/getScopeTestPathsRecursively";
 export * from "./util/typeUtils";
 export * from "./ide/types/hatStyles.types";
 export * from "./errors";
