+++ packages/common/src/types/ScopeTestRecorderStorage.ts
@@ -1,0 +1,8 @@
+export interface ScopeTestRecorderStorage {
+  getTestedScopeFacets(languageId: string): Set<string>;
+  saveScopeFacetTest(
+    languageId: string,
+    facet: string,
+    content: string,
+  ): Promise<void>;
+}
