+++ packages/cursorless-engine/src/ScopeTestRecorder.ts
@@ -1,20 +1,18 @@
 import {
   ScopeSupportFacetLevel,
-  getScopeTestPathsRecursively,
-  getScopeTestsDirPath,
-  groupBy,
   languageScopeSupport,
   scopeSupportFacetInfos,
   showInfo,
   type IDE,
   type ScopeSupportFacet,
+  type ScopeTestRecorderStorage,
 } from "@cursorless/common";
-import * as fs from "node:fs";
-import * as fsPromises from "node:fs/promises";
-import * as path from "pathe";
 
 export class ScopeTestRecorder {
-  constructor(private ide: IDE) {
+  constructor(
+    private ide: IDE,
+    private storage: ScopeTestRecorderStorage,
+  ) {
     this.showUnimplementedFacets = this.showUnimplementedFacets.bind(this);
     this.saveActiveDocument = this.saveActiveDocument.bind(this);
   }
@@ -27,7 +25,8 @@ export class ScopeTestRecorder {
     }
 
     const supportedScopeFacets = getSupportedScopeFacets(languageId);
-    const existingScopeTestFacets = getExistingScopeFacetTest(languageId);
+    const existingScopeTestFacets =
+      this.storage.getTestedScopeFacets(languageId);
 
     const missingScopeFacets = supportedScopeFacets.filter(
       (facet) => !existingScopeTestFacets.has(facet),
@@ -83,20 +82,10 @@ export class ScopeTestRecorder {
       facetsToAdd.push({ facet, content });
     }
 
-    const langDirectory = path.join(getScopeTestsDirPath(), languageId);
-
-    await fsPromises.mkdir(langDirectory, { recursive: true });
-
     for (const { facet, content } of facetsToAdd) {
       const fullContent = `${content}---\n`;
-      let filePath = path.join(langDirectory, `${facet}.scope`);
-      let i = 2;
 
-      while (fs.existsSync(filePath)) {
-        filePath = path.join(langDirectory, `${facet}${i++}.scope`);
-      }
-
-      await fsPromises.writeFile(filePath, fullContent, "utf-8");
+      this.storage.saveScopeFacetTest(languageId, facet, fullContent);
     }
 
     await showInfo(
@@ -127,12 +116,4 @@ function getSupportedScopeFacets(languageId: string): ScopeSupportFacet[] {
   return scopeFacets.filter(
     (facet) => scopeSupport[facet] === ScopeSupportFacetLevel.supported,
   );
-}
-
-function getExistingScopeFacetTest(languageId: string): Set<string> {
-  const testPaths = getScopeTestPathsRecursively();
-  const languages = groupBy(testPaths, (test) => test.languageId);
-  const testPathsForLanguage = languages.get(languageId) ?? [];
-  const facets = testPathsForLanguage.map((test) => test.facet);
-  return new Set(facets);
 }
