+++ packages/cursorless-engine/src/disabledComponents/DisabledTalonSpokenForms.ts
@@ -1,7 +1,7 @@
 import type {
   SpokenFormEntry,
   TalonSpokenForms,
-} from "../scopeProviders/TalonSpokenForms";
+} from "../../../common/src/types/TalonSpokenForms";
 
 export class DisabledTalonSpokenForms implements TalonSpokenForms {
   getSpokenFormEntries(): Promise<SpokenFormEntry[]> {
