+++ packages/cursorless-engine/src/generateSpokenForm/CustomSpokenFormGeneratorImpl.ts
@@ -8,7 +8,7 @@ import {
 import { SpokenFormGenerator } from "./generateSpokenForm";
 import { CustomSpokenFormGenerator } from "../api/CursorlessEngineApi";
 import { CustomSpokenForms } from "../spokenForms/CustomSpokenForms";
-import { TalonSpokenForms } from "../scopeProviders/TalonSpokenForms";
+import { TalonSpokenForms } from "../../../common/src/types/TalonSpokenForms";
 
 /**
  * Simple facade that combines the {@link CustomSpokenForms} and
