+++ packages/cursorless-engine/src/generateSpokenForm/generateSpokenForm.test.ts
@@ -1,17 +1,17 @@
 import {
   TestCaseFixtureLegacy,
-  getRecordedTestPaths,
   serializeTestFixture,
   shouldUpdateFixtures,
 } from "@cursorless/common";
+import { getRecordedTestPaths } from "@cursorless/node-common";
 import * as yaml from "js-yaml";
 import assert from "node:assert";
 import { promises as fsp } from "node:fs";
-import { canonicalizeAndValidateCommand } from "../core/commandVersionUpgrades/canonicalizeAndValidateCommand";
-import { getHatMapCommand } from "./getHatMapCommand";
 import { SpokenFormGenerator } from ".";
-import { defaultSpokenFormInfoMap } from "../spokenForms/defaultSpokenFormMap";
+import { canonicalizeAndValidateCommand } from "../core/commandVersionUpgrades/canonicalizeAndValidateCommand";
 import { SpokenFormMap, mapSpokenForms } from "../spokenForms/SpokenFormMap";
+import { defaultSpokenFormInfoMap } from "../spokenForms/defaultSpokenFormMap";
+import { getHatMapCommand } from "./getHatMapCommand";
 
 /**
  * A spoken form map to use for testing. Just uses default spoken forms, but
