+++ packages/cursorless-engine/src/index.ts
@@ -11,7 +11,7 @@ export * from "./CommandHistory";
 export * from "./CommandHistoryAnalyzer";
 export * from "./util/grammarHelpers";
 export * from "./ScopeTestRecorder";
-export * from "./scopeProviders/TalonSpokenForms";
+export * from "../../common/src/types/TalonSpokenForms";
 export * from "./core/Snippets";
 export * from "./core/mergeSnippets";
 export * from "./util/object";
