+++ packages/cursorless-engine/src/test/spokenForms.talon.test.ts
@@ -3,19 +3,19 @@ import {
   CommandLatest,
   TestCaseFixtureLegacy,
   asyncSafety,
-  getRecordedTestPaths,
 } from "@cursorless/common";
+import { getRecordedTestPaths } from "@cursorless/node-common";
 import assert from "assert";
 import * as yaml from "js-yaml";
 import { promises as fsp } from "node:fs";
 import { canonicalizeAndValidateCommand } from "../core/commandVersionUpgrades/canonicalizeAndValidateCommand";
 import { getHatMapCommand } from "../generateSpokenForm/getHatMapCommand";
 import { TalonRepl } from "../testUtil/TalonRepl";
+import { communitySnippetsSpokenFormsFixture } from "./fixtures/communitySnippets.fixture";
+import { multiActionFixture } from "./fixtures/multiAction.fixture";
+import { SpokenFormTestOpts } from "./fixtures/spokenFormTest";
 import { synonymousSpokenFormsFixture } from "./fixtures/synonymousSpokenForms.fixture";
 import { talonApiFixture } from "./fixtures/talonApi.fixture";
-import { multiActionFixture } from "./fixtures/multiAction.fixture";
-import { communitySnippetsSpokenFormsFixture } from "./fixtures/communitySnippets.fixture";
-import { SpokenFormTestOpts } from "./fixtures/spokenFormTest";
 
 suite("Talon spoken forms", async function () {
   const repl = new TalonRepl();
