+++ packages/cursorless-vscode-e2e/src/suite/followLink.vscode.test.ts
@@ -4,7 +4,7 @@ import * as os from "os";
 import * as vscode from "vscode";
 import { endToEndTestSetup } from "../endToEndTestSetup";
 import { runCursorlessCommand } from "@cursorless/vscode-common";
-import { getFixturePath } from "@cursorless/common";
+import { getFixturePath } from "@cursorless/node-common";
 
 suite("followLink", async function () {
   endToEndTestSetup(this);
