+++ packages/cursorless-vscode-e2e/src/suite/scopes.vscode.test.ts
@@ -1,6 +1,5 @@
 import {
   asyncSafety,
-  getScopeTestPathsRecursively,
   languageScopeSupport,
   ScopeSupportFacet,
   scopeSupportFacetInfos,
@@ -10,6 +9,7 @@ import {
   TextualScopeSupportFacet,
   textualScopeSupportFacetInfos,
 } from "@cursorless/common";
+import { getScopeTestPathsRecursively } from "@cursorless/node-common";
 import { getCursorlessApi, openNewEditor } from "@cursorless/vscode-common";
 import { assert } from "chai";
 import { groupBy, uniq } from "lodash-es";
