+++ packages/cursorless-vscode-e2e/src/suite/testCaseRecorder.vscode.test.ts
@@ -1,10 +1,12 @@
 import {
-  getFixturePath,
-  getRecordedTestsDirPath,
   HatTokenMap,
   LATEST_VERSION,
   SimpleActionName,
 } from "@cursorless/common";
+import {
+  getFixturePath,
+  getRecordedTestsDirPath,
+} from "@cursorless/node-common";
 import {
   getCursorlessApi,
   openNewEditor,
@@ -12,7 +14,7 @@ import {
 } from "@cursorless/vscode-common";
 import { assert } from "chai";
 import * as crypto from "crypto";
-import { mkdir, readdir, readFile, rm } from "fs/promises";
+import { mkdir, readFile, readdir, rm } from "fs/promises";
 import * as os from "os";
 import * as path from "pathe";
 import * as vscode from "vscode";
