+++ packages/node-common/package.json
@@ -25,9 +25,11 @@
   },
   "dependencies": {
     "@cursorless/common": "workspace:*",
-    "@cursorless/cursorless-engine": "workspace:*",
     "glob": "^10.3.10",
-    "pathe": "^1.1.2"
+    "pathe": "^1.1.2",
+    "lodash-es": "^4.17.21"
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "@types/lodash-es": "4.17.0"
+  }
 }
