+++ packages/node-common/src/getScopeTestPathsRecursively.ts
