+++ packages/node-common/src/index.ts
@@ -1,4 +1,8 @@
 export * from "./FileSystemCommandHistoryStorage";
 export * from "./FileSystemLanguageDefinitionsProvider";
+export * from "./FileSystemScopeTestRecorderStorage";
 export * from "./FileSystemTalonSpokenForms";
 export * from "./nodeGetRunMode";
+export * from "./getCursorlessRepoRoot";
+export * from "./getFixturePaths";
+export * from "./getScopeTestPathsRecursively";
