+++ packages/node-common/src/walkSync.ts
@@ -1,4 +1,4 @@
-import * as path from "pathe";
+import * as path from "path";
 import { readdirSync } from "fs";
 
 /**
