+++ packages/node-test/package.json
@@ -1,37 +1,36 @@
 {
-  "name": "@cursorless/cursorless-cheatsheet",
-  "version": "0.1.0",
-  "description": "Cursorless cheatsheet implementation",
-  "type": "module",
+  "name": "@cursorless/node-test",
+  "version": "1.0.0",
+  "description": "Package containing node test utilities",
   "main": "./out/index.js",
-  "types": "./out/index.d.ts",
   "scripts": {
     "compile:tsc": "tsc --build",
     "compile:esbuild": "esbuild ./src/index.ts --sourcemap --format=esm --bundle --packages=external --outfile=./out/index.js",
     "compile": "pnpm compile:tsc && pnpm compile:esbuild",
     "watch:tsc": "pnpm compile:tsc --watch",
     "watch:esbuild": "pnpm compile:esbuild --watch",
-    "watch": "pnpm run --filter @cursorless/cursorless-cheatsheet --parallel '/^watch:.*/'",
+    "watch": "pnpm run --filter @cursorless/node-test --parallel '/^watch:.*/'",
     "clean": "rm -rf ./out tsconfig.tsbuildinfo ./dist ./build"
   },
   "keywords": [],
   "author": "",
   "license": "MIT",
+  "type": "module",
+  "types": "./out/index.d.ts",
   "exports": {
     ".": {
       "cursorless:bundler": "./src/index.ts",
       "default": "./out/index.js"
     }
   },
-  "dependencies": {
+  "dependencies": {},
+  "devDependencies": {
     "@cursorless/common": "workspace:*",
     "@cursorless/node-common": "workspace:*",
-    "immer": "^10.0.4",
-    "lodash-es": "^4.17.21",
-    "node-html-parser": "^6.1.12",
-    "pathe": "^1.1.2"
-  },
-  "devDependencies": {
-    "@types/lodash-es": "4.17.0"
+    "@types/chai": "^4.3.14",
+    "@types/js-yaml": "^4.0.9",
+    "@types/lodash-es": "4.17.0",
+    "js-yaml": "^4.1.0",
+    "lodash-es": "^4.17.21"
   }
 }
