+++ packages/test-harness/src/testSubset.ts
@@ -1,6 +1,6 @@
 import * as fs from "fs";
 import * as path from "pathe";
-import { getCursorlessRepoRoot } from "@cursorless/common";
+import { getCursorlessRepoRoot } from "@cursorless/node-common";
 
 /**
  * Returns the grep string to pass to Mocha when running a subset of tests.
