+++ packages/test-harness/tsconfig.json
@@ -8,6 +8,9 @@
   "references": [
     {
       "path": "../common"
+    },
+    {
+      "path": "../node-common"
     }
   ]
 }
