+++ packages/vscode-common/package.json
@@ -16,7 +16,8 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@cursorless/common": "workspace:*"
+    "@cursorless/common": "workspace:*",
+    "@cursorless/node-test": "workspace:*"
   },
   "devDependencies": {
     "@types/vscode": "1.75.1"
