+++ packages/vscode-common/tsconfig.json
@@ -9,6 +9,9 @@
   "references": [
     {
       "path": "../common"
+    },
+    {
+      "path": "../node-common"
     }
   ]
 }
